+++ packages/services/indexer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsequence/indexer
 
+## 3.0.4
+
+### Patch Changes
+
+- id-token login support
+
+## 3.0.3
+
+### Patch Changes
+
+- 3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- allow native self transfer
+
 ## 3.0.1
 
 ### Patch Changes